chore: update MSRV to 1.88 and fix cross-platform CI issues
- Bump MSRV from 1.85 to 1.88 across all config files
- Re-enable ssh feature on Windows (pageant let chains now supported)
- Use latest cargo-deny (supports CVSS 4.0 advisory format)
- Fix macOS ioctl type mismatch: cast TIOCSCTTY and TIOCSWINSZ to c_ulong
- Add PtySystem trait import to windows_pty_test.rs example

This resolves all CI failures related to:
- pageant 0.2.0 requiring Rust 1.88 for let chains
- cargo-deny 0.18.9 requiring Rust 1.88
- CVSS 4.0 format in RustSec advisory database
- macOS type mismatch between u32 ioctl constants and u64 c_ulong

+++ crates/rust-expect/src/backend/pty.rs
@@ -236,7 +236,8 @@ impl PtySpawner {
                 unsafe {
                     libc::close(master_fd);
                     libc::setsid();
-                    libc::ioctl(slave_fd, libc::TIOCSCTTY, 0);
+                    // Cast TIOCSCTTY to c_ulong for macOS compatibility (u32 -> u64)
+                    libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0);
 
                     libc::dup2(slave_fd, 0);
                     libc::dup2(slave_fd, 1);
@@ -389,7 +390,9 @@ impl PtyHandle {
         // SAFETY: master_fd is a valid PTY file descriptor stored in self.
         // TIOCSWINSZ is a valid ioctl command for PTYs that sets the window size.
         // winsize is a valid pointer to a properly initialized struct on the stack.
-        let result = unsafe { libc::ioctl(self.master_fd, libc::TIOCSWINSZ, &winsize) };
+        // Cast to c_ulong for macOS compatibility (u32 -> u64).
+        let result =
+            unsafe { libc::ioctl(self.master_fd, libc::TIOCSWINSZ as libc::c_ulong, &winsize) };
 
         if result != 0 {
             Err(ExpectError::Io(io::Error::last_os_error()))
@@ -560,7 +563,14 @@ impl AsyncPty {
         };
 
         // SAFETY: The fd is valid and TIOCSWINSZ is a valid ioctl for PTYs.
-        let result = unsafe { libc::ioctl(*self.inner.get_ref(), libc::TIOCSWINSZ, &winsize) };
+        // Cast to c_ulong for macOS compatibility (u32 -> u64).
+        let result = unsafe {
+            libc::ioctl(
+                *self.inner.get_ref(),
+                libc::TIOCSWINSZ as libc::c_ulong,
+                &winsize,
+            )
+        };
 
         if result != 0 {
             Err(ExpectError::Io(io::Error::last_os_error()))
